LayoutTests/http/tests/security/contentSecurityPolicy/sandbox-report-only.html

<!DOCTYPE html>
<iframe src="resources/sandbox.php?report-only=1"></iframe>
<script>
if (window.testRunner)
    testRunner.dumpAsText();

window.onload = function() {
    var frame = document.getElementsByTagName("iframe")[0];
    var secret = frame.contentWindow.secret;
    if (secret)
        alert("PASS: Iframe was not in a unique origin");
    else
        alert("FAIL: Iframe was in a unique origin");
};
</script>
