third_party/blink/web_tests/platform/win11-arm64/http/tests/security/contentSecurityPolicy/sandbox-report-only-expected.txt

CONSOLE MESSAGE: Script executed in iframe.
ALERT: PASS: Iframe was not in a unique origin
